split off web and cli
